tests: put tests for scikit-learn and scikit-image into separate files
