+++ src/_pyinstaller_hooks_contrib/tests/test_libraries.py
@@ -489,43 +489,6 @@ def test_pydivert(pyi_builder):
         """)
 
 
-@pytest.mark.slow
-@importorskip('skimage')
-@pytest.mark.skipif(not is_module_satisfies('scikit_image >= 0.16'),
-                    reason='The test supports only scikit-image >= 0.16.')
-@pytest.mark.parametrize('submodule', [
-    'color', 'data', 'draw', 'exposure', 'feature', 'filters', 'future',
-    'graph', 'io', 'measure', 'metrics', 'morphology', 'registration',
-    'restoration', 'segmentation', 'transform', 'util'
-])
-def test_skimage(pyi_builder, submodule):
-    pyi_builder.test_source("""
-        import skimage.{0}
-        """.format(submodule))
-
-
-@pytest.mark.slow
-@importorskip('sklearn')
-@pytest.mark.skipif(not is_module_satisfies('scikit_learn >= 0.21'),
-                    reason='The test supports only scikit-learn >= 0.21.')
-@pytest.mark.parametrize('submodule', [
-    'calibration', 'cluster', 'covariance', 'cross_decomposition',
-    'datasets', 'decomposition', 'dummy', 'ensemble', 'exceptions',
-    'experimental', 'externals', 'feature_extraction',
-    'feature_selection', 'gaussian_process', 'inspection',
-    'isotonic', 'kernel_approximation', 'kernel_ridge',
-    'linear_model', 'manifold', 'metrics', 'mixture',
-    'model_selection', 'multiclass', 'multioutput',
-    'naive_bayes', 'neighbors', 'neural_network', 'pipeline',
-    'preprocessing', 'random_projection', 'semi_supervised',
-    'svm', 'tree', 'discriminant_analysis', 'impute', 'compose'
-])
-def test_sklearn(pyi_builder, submodule):
-    pyi_builder.test_source("""
-        import sklearn.{0}
-        """.format(submodule))
-
-
 @importorskip('statsmodels')
 @pytest.mark.skipif(not is_module_satisfies('statsmodels >= 0.12'),
                     reason='This has only been tested with statsmodels >= 0.12.')
